Fix potential broken deactivate bug
xscreensaver-command waits for a second when activating to work around
keyup events causing deactivation. If a deactivate command is sent during
this time, the screensaver can activate after deactivate happens which
causes confusion for a caller keeping track. Here we add a ref counter
to keep track and discard events when the counter is out of sync.